fix: address off by one error causing line gaps
one of the issues around porting smalltalk code or using it as a ways
to understand an algorithm or implementation is that it's very easy to
forget that smalltalk is 1-indexed.

as a result of that whenever `doodle` was trying to draw a line
between points separated by 1 pixel - for example, when moving the
mouse very slowly - nothing would draw because `range(1,1)` is empty!

this commit fixes that and makes some changes to `doodle` to avoid
crashing.

there's still a couple of `bitblt` bugs:
- there's some gaps when draving up and to the right with small brush.
  - not sure about this.
- out of bounds errors when going beyond the destination's width.
  - i suspect a clipping issue here.